review processed + added fixes to dockerfiles and makefile for new v6 registry
